requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
faiss-cpu>=1.8.0
fastembed>=0.7.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.1.0
